composition_polynomials file
